Trasladé todo el código a main.py para desarrollar con claridad. Después separaré los módulos.
