--- semana_10/proyecto_semana_10/actions/act2_print_student_info.py
@@ -1,5 +1,0 @@
-
-
-from students_info import students_list
-
-print(students_list)
